fix(api): validatecq enabled function shared
Ref. metriport/metriport-internal#1350
